# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PostMom Server API
  version: 1.0.0
  description: Express backend API for chat and community features.
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Server is healthy
  /api/community/posts:
    get:
      summary: Get community posts
      responses:
        '200':
          description: Community post list
    post:
      summary: Create community post
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Post created
        '401':
          description: Unauthorized
  /api/community/posts/{id}:
    get:
      summary: Get one community post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post detail
        '404':
          description: Not found
    patch:
      summary: Update community post
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Post updated
    delete:
      summary: Delete community post
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Post deleted
  /api/community/posts/{id}/comments:
    get:
      summary: Get comments by post id
      responses:
        '200':
          description: Comment list
    post:
      summary: Create comment
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Comment created
  /api/community/comments/{commentId}:
    delete:
      summary: Delete comment
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Comment deleted
  /api/community/posts/{id}/likes:
    post:
      summary: Like post
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Liked
    delete:
      summary: Unlike post
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unliked
  /api/community/likes/me:
    get:
      summary: Get my liked post ids
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of liked post ids
  /api/sessions:
    get:
      summary: Get chat sessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Chat sessions
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT